fix(ci): repair plugin prerelease validation
